fix: resolve cross-platform build issues and refactor algorithms
- Fix transitive include errors on Linux/Clang by explicitly including missing headers (`<cstdint>`, `<expected>`, `<cstddef>`, `<string_view>`).
- Refactor `std::lower_bound` to C++20 `std::ranges::lower_bound` using projections to resolve `strict_weak_order` constraint failures.
- Resolve PIMPL incomplete type error by moving `Json` move constructor and assignment operator definitions to the source file.
- Fix `undefined symbol: main` in example files by renaming functions to `int main()`.
- Add comprehensive Markdown documentation including Example Pack and C++20 Ranges fundamentals.

+++ example/example_array.cpp
@@ -27,7 +27,7 @@ int main() {
         std::println("zuu punya {} bahasa pemrograman favorit:", bahasa.size());
 
         // Iterasi array secara aman
-        for (size_t i = 0; i < bahasa.size(); ++i) {
+        for (unsigned long long i = 0; i < bahasa.size(); ++i) {
             std::println("{}. {}", i + 1, bahasa[i].get_string());
         }
     }
