+++ src/models/storage.cpp
@@ -14,22 +14,23 @@
 namespace zuu::models {
 
 Storage::Storage(models::Hint<Token> hint) noexcept {
-    const size_t max_strings = hint.string_count_;
-    const size_t max_arrays = hint.array_count_;
-    const size_t max_objects = hint.object_count_;
-    const size_t max_elements = hint.comma_count_ + hint.array_count_ + hint.object_count_;
+    const unsigned long long max_strings = hint.string_count_;
+    const unsigned long long max_arrays = hint.array_count_;
+    const unsigned long long max_objects = hint.object_count_;
+    const unsigned long long max_elements =
+        hint.comma_count_ + hint.array_count_ + hint.object_count_;
 
-    const size_t strings_bytes = max_strings * sizeof(std::string_view);
-    const size_t array_elem_bytes = max_elements * sizeof(JsonValue);
-    const size_t arrays_bytes = max_arrays * sizeof(std::pair<uint32_t, uint32_t>);
-    const size_t obj_elem_bytes = max_elements * sizeof(JsonMember);
-    const size_t objects_bytes = max_objects * sizeof(std::pair<uint32_t, uint32_t>);
+    const unsigned long long strings_bytes = max_strings * sizeof(std::string_view);
+    const unsigned long long array_elem_bytes = max_elements * sizeof(JsonValue);
+    const unsigned long long arrays_bytes = max_arrays * sizeof(std::pair<unsigned, unsigned>);
+    const unsigned long long obj_elem_bytes = max_elements * sizeof(JsonMember);
+    const unsigned long long objects_bytes = max_objects * sizeof(std::pair<unsigned, unsigned>);
 
     // Alokasi ruang untuk unescaped characters (beserta margin aman 10%)
-    const size_t str_buf_bytes = hint.string_escape_bytes_;
+    const unsigned long long str_buf_bytes = hint.string_escape_bytes_;
 
-    const size_t total_bytes = strings_bytes + array_elem_bytes + arrays_bytes + obj_elem_bytes +
-                               objects_bytes + str_buf_bytes;
+    const unsigned long long total_bytes = strings_bytes + array_elem_bytes + arrays_bytes +
+                                           obj_elem_bytes + objects_bytes + str_buf_bytes;
 
     if (total_bytes > 0) {
         arena_ = std::make_unique<std::byte[]>(total_bytes);
@@ -42,13 +43,13 @@ Storage::Storage(models::Hint<Token> hint) noexcept {
         array_elements_ = reinterpret_cast<JsonValue*>(ptr);
         ptr += array_elem_bytes;
 
-        arrays_ = reinterpret_cast<std::pair<uint32_t, uint32_t>*>(ptr);
+        arrays_ = reinterpret_cast<std::pair<unsigned, unsigned>*>(ptr);
         ptr += arrays_bytes;
 
         object_elements_ = reinterpret_cast<JsonMember*>(ptr);
         ptr += obj_elem_bytes;
 
-        objects_ = reinterpret_cast<std::pair<uint32_t, uint32_t>*>(ptr);
+        objects_ = reinterpret_cast<std::pair<unsigned, unsigned>*>(ptr);
         ptr += objects_bytes;
 
         // Ditempatkan paling akhir agar tidak merusak byte alignment dari tipe data lain di atasnya
@@ -71,18 +72,18 @@ const JsonValue& Storage::root() const noexcept {
 
 // ── BUMP ALLOCATOR OPERATIONS ───────────────────────────────────────────────
 
-size_t Storage::commitString(std::string_view value) noexcept {
+unsigned long long Storage::commitString(std::string_view value) noexcept {
     strings_[strings_size_] = value;
     return strings_size_++;
 }
 
-char* Storage::allocateStringBuffer(size_t length) noexcept {
+char* Storage::allocateStringBuffer(unsigned long long length) noexcept {
     char* allocated = string_buffer_ + string_buffer_size_;
-    string_buffer_size_ += static_cast<uint32_t>(length);
+    string_buffer_size_ += static_cast<unsigned>(length);
     return allocated;
 }
 
-size_t Storage::getArrayOffset() const noexcept {
+unsigned long long Storage::getArrayOffset() const noexcept {
     return array_elements_size_;
 }
 
@@ -90,13 +91,13 @@ void Storage::pushArrayElement(const JsonValue& val) noexcept {
     array_elements_[array_elements_size_++] = val;
 }
 
-size_t Storage::sealArray(size_t start_offset) noexcept {
-    const auto size = static_cast<uint32_t>(array_elements_size_ - start_offset);
-    arrays_[arrays_size_] = {static_cast<uint32_t>(start_offset), size};
+unsigned long long Storage::sealArray(unsigned long long start_offset) noexcept {
+    const auto size = static_cast<unsigned>(array_elements_size_ - start_offset);
+    arrays_[arrays_size_] = {static_cast<unsigned>(start_offset), size};
     return arrays_size_++;
 }
 
-size_t Storage::getObjectOffset() const noexcept {
+unsigned long long Storage::getObjectOffset() const noexcept {
     return object_elements_size_;
 }
 
@@ -104,8 +105,8 @@ void Storage::pushObjectMember(const JsonMember& member) noexcept {
     object_elements_[object_elements_size_++] = member;
 }
 
-size_t Storage::sealObject(size_t start_offset) noexcept {
-    const auto size = static_cast<uint32_t>(object_elements_size_ - start_offset);
+unsigned long long Storage::sealObject(unsigned long long start_offset) noexcept {
+    const auto size = static_cast<unsigned>(object_elements_size_ - start_offset);
     if (size > 1) {
         std::sort(object_elements_ + start_offset,
                   object_elements_ + object_elements_size_,
@@ -113,23 +114,23 @@ size_t Storage::sealObject(size_t start_offset) noexcept {
                       return strings_[a.key_index_] < strings_[b.key_index_];
                   });
     }
-    objects_[objects_size_] = {static_cast<uint32_t>(start_offset), size};
+    objects_[objects_size_] = {static_cast<unsigned>(start_offset), size};
     return objects_size_++;
 }
 
 // ─────────────────────────────────────────────────────────────────────────────
 
-Storage::JsonArray Storage::array(size_t index) const noexcept {
+Storage::JsonArray Storage::array(unsigned long long index) const noexcept {
     const auto& [offset, size] = arrays_[index];
     return {array_elements_ + offset, size};
 }
 
-Storage::JsonObject Storage::object(size_t index) const noexcept {
+Storage::JsonObject Storage::object(unsigned long long index) const noexcept {
     const auto& [offset, size] = objects_[index];
     return {object_elements_ + offset, size};
 }
 
-std::string_view Storage::string(size_t index) const noexcept {
+std::string_view Storage::string(unsigned long long index) const noexcept {
     return strings_[index];
 }
